requirements.txt
# Core packages
numpy==1.26.4
Pillow==10.4.0
matplotlib==3.9.2
scikit-image==0.24.0

# Web framework
Flask==3.0.3
flask-cors==4.0.0
werkzeug==3.0.1
python-dotenv>=1.0.0
requests>=2.31.0

# Machine learning / AI
tensorflow==2.19.0
torch==2.1.0
torchvision==0.16.0
joblib>=1.3.0

# Image Processing
opencv-python==4.8.1.78

# Explainability
shap==0.46.0
lime==0.2.0.1
grad-cam>=1.4.8

# Google Gemini AI
google-generativeai==0.8.3

# Utilities
scikit-learn==1.3.2

# LLM + RAG
chromadb>=0.5.0
langchain>=0.2.0
langchain-chroma>=0.1.0
langchain-google-genai>=1.0.0
pypdf>=4.0.0
sentence-transformers>=2.7.0
ctransformers>=0.2.27
